requirements.txt
jinja2
grafana_client
pyyaml